Spicy.Math

Description

This module defines basic algebraic operations used throughout the program. Numerical heavy and most other operations are implemented using Accelerate, to provide parallel operations. Note that all runQ provided functions must be typed without typeclasses but by concrete types.

Documentation

(∩) :: Eq a => [a] -> [a] -> [a]

Intersection (subset) of two lists a and b.

(<.>) :: Vector Double -> Vector Double -> Double

Dot product of two Vectors.