Spicy.Math.Helper

Description

Definitions of auxiliary functions to deal with the GHC bug, making it necessary to have all runQ expressions as untyped slices. Functions in this module are not meant for direct use, but wrapped in the Math module.

Documentation

vvP :: Acc (Vector Double) -> Acc (Vector Double) -> Acc (Scalar Double)

Vector vector dot product.

vmP :: Acc (Vector Double) -> Acc (Matrix Double) -> Acc (Vector Double)

Vector matrix product.

mvP :: Acc (Matrix Double) -> Acc (Vector Double) -> Acc (Vector Double)

Matrix vector product.

mmP :: Acc (Matrix Double) -> Acc (Matrix Double) -> Acc (Matrix Double)

Matrix matrix product.