Spicy.Math.Internal

Description

Definitions of auxiliary functions to deal with the GHC bug, making it necessary to have all runQ expressions as untyped slices. Functions in this module are not meant for direct use, but wrapped in the Math module.

Documentation

getCoordinates1 :: Strat -> Molecule -> Vector Double

Get the Atom _atom_Coordinates from a Molecule and convert to a plain Vector. This is therefore basically a concatenation of all cartesian coordinates

getCoordinates2 :: Strat -> Molecule -> Vector Double