Spicy.MolWriter

Description

A module which converts the internal Molecule representation to a string, which is a common chemical file format, that can be read by Avogadro, VMD, OpenBabel etc.. The writers are not fool proof with respect to force field types, which should always be remembered when usings its results.

Documentation

writeXYZ :: Molecule -> Text

Write a Molden XYZ file from a molecule. This format ignores all deep level layers of a molecule.

writeTXYZ :: Molecule -> Text

Write a Tinker XYZ from a Molecule. The writer trusts the _atom_FFType to be correct (if set) and will simply write them out. Therefore it is possible, that wrong atom types can be written. If they are not set, the writer will simply equalise all atom types to 0, which is OK for visualisation but obviously not for MM.

This format ingores all deeper level layers of a molecule.

writeMOL2 :: Molecule -> Text

Write a simplified .mol2 file (Tripos SYBYL) from a Molecule, containing the atoms, connectivities (single bonds only) and partial charges. The writer is not fool proof and will happily accept any _atom_FFType that is supplied, even if it is not a TRIPOS SYBYL atom type. This can lead to mol2 files that have correct topology and geometry, but visualisation programs wont be able to assign correct elements.

writeSpicy :: Molecule -> String

Write Spicy format, which is a custom format (not stable yet), containing all informations, that are internally used to represent a molecule.

data Molecule

A molecule, which might be the whole system, an ONIOM layer or a fragment of the system, each containing possibly even higher layers for ONIOM or fragments. Stores all associated informations of a layer.

Starting from a top level molecule, all atoms and bonds of the system are expected to be in the in this top layer (except pseudoatoms of deeper layers). Therefore if atoms are in a deeper layers of the recursion, their information is not used to describe a higher layer. Instead, all atoms of deeper layers (except pseudoatoms) must be also replicated in a higher layer.

Two possible types of index countings are possible:

Global: All atoms in the system (which should be present at the top layer molecule) except pseudoatoms are counted in the top layer. The fragments might therefore not start at atom with index 0, but depending on where they are in a higher level with higher numbers. Pseudoatoms are counted after all other atoms are counted.

Individual: Counting happens for each molecule individually. In this counting scheme, pseudoatoms do not need any special treatment.

Constructors

Molecule

Fields

_molecule_Label :: String

Comment or identifier of a molecule. Can be empty.

_molecule_Atoms :: Vector Atom

A Vector of Atoms.

_molecule_Bonds :: IntMap IntSet

An IntMap, mapping the index of an Atom in the Molecule to the indices of all Atoms, to which it binds.

_molecule_SubMol :: Vector Molecule

A Molecule might contain other molecules. These might be fragments or higher level ONIOM layers.

_molecule_IndType :: IndType

Which counting scheme for atoms is used in this representation. Must be same for all layers.

_molecule_Energy :: Maybe Double

An energy, that might have been calculated.

_molecule_Gradient :: Maybe (Vector Double)

A gradient, that might have been calculated.

_molecule_Hessian :: Maybe (UMatrix Double)

A hessian, that might have been calculated.