Spicy.Molecule.Util

Description

This module provides functions to manipulate basic data structures of Molecules, such as indexing.

Documentation

reIndexBonds

Arguments

:: IntMap Int
Substitution patterns with original as key and new as Value.

-> IntMap IntSet
Original _molecule_Bonds styled IntMap.

-> Either String (IntMap IntSet)
New bonds after replacements. Results in a Left String as error, if not all old indices can be mapped to new indices.

Calculate new indices for Molecules in the Individual scheme, where each fragment is counted by itself without awareness for other molecules. This relies on sane bonds in the Molecule. If somewhere bonds are defined to non existing atoms, these bonds will be lost.

Reindex a single layer of a molecule (topmost) but ignore its deeper layers. Reindexing is done based on a replacement map, which must not have any holes. So Every index that appears in the molecule must also appear in the replacement list, even if this maps to the same value. Otherwise a Left error String will result.

Substitute all indices in the bonding style IntMap by replacements given in a IntMap Int. If keys of the old bonding map do not appear in the replacement map, a Left error String will be given. Therefore the replacement map must be complete and no holes are allowed.

deleteIntMapByKeys

Arguments

:: [Key]
List of Keys to be removed from an IntMap

-> IntMap a
Original IntMap, where Keys shall be deleted.

-> IntMap a
IntMap after deletion.

Given a list of Keys, delete all Keys in the list from the original IntMap.

deleteIntSetByKeys

Arguments

:: [Key]
List of Keys to be removed from an IntSet.

-> IntSet
IntSet from which to remove Keys.

-> IntSet
IntSet after deletion of Keys.

Given a list of Keys, delete all Keys in the list from the original IntSet.

updateKeys

Arguments

:: IntMap Int
IntMap of replacements. Old Key mapped to a new Key.

-> IntMap a
Original IntMap in which to replace the Keys.

-> IntMap a
IntMap after replacements.

Given a replacement IntMap (Key is the old value, value is the new Key), replace the Keys from the old IntMap with the ones from the replacement IntMap. If a key cannot be found in the replacement IntMap, the original Key of the old IntMap will be kept.

updateIntSet

Arguments

:: IntMap Int
IntMap of replacements. Old Key mapped to a new Key.

-> IntSet
Original IntSet in which to replace Keys

-> IntSet
IntSet after replacements done.

Given a replacement IntMap, replace all values in an original IntSet by values from a replacement IntMap. Values from the IntSet, that have no Key in the IntMap will remain unchanged.