Spicy.Types

Description

Spicy.Types contains the definition of all classes and data types, that are used in Spicy. Mainly it takes care of the description of molecules (structure, topology, potential energy surface, ...) and computations on molecules in different software packages.

Documentation

class NiceShow a where

Have a class for the printing of the calculation niveau, which allows to produce a very readable output.

Methods

niceShow

Arguments

:: a

-> String
Printing the isolated object

niceComplex

Arguments

:: a

-> String
Printing the object in the complex form, where everything is meant to be printed at once as overview

data Strat

Use serial or parallel processing for large data structures. This helps deciding on a per use base, if to evaluate in serial or parallel, to avoid nested parallelism. Every method employing parallel operations by Control.Parallel.Strategies should provide this switch in Spicy.

Constructors

Serial

Parallel

data IndType

Indexing type for recursive molecule structures. Count all layers with a global index range (Global) or by fragment specific index range (Individual).

Constructors

Global

Individual

type UMatrix a = Array (Int, Int) a

Alias for an unboxed matrix from the Array library. This type can easily be converted to Accelerate's array type.

Types to describe a molecule from a computational geometry point of view. The representation of a molecule aims to be as descriptive as unambigous as possible and capture all quantities relevant for molecular dynamics and geometry optimisations (energies, gradients, hessians, force fiekd types, coordinates, connectivities). Electronic properties, such as dipoles, orbitals, the wavefunction and so on, will be ignored, as Spicy is not aiming to be an analysis program, but a interface to create multilayer calculations. Analysis of the results is up to postprocessing.

data Element

All chemical elements. Have them very clear because force fields and pdb names may interfer and are just arbitrary strings.

Constructors

H

He

Li

Be

B

C

N

O

F

Ne

Na

Mg

Al

Si

P

S

Cl

Ar

K

Ca

Sc

Ti

V

Cr

Mn

Fe

Co

Ni

Cu

Zn

Ga

Ge

As

Se

Br

Kr

Rb

Sr

Y

Zr

Nb

Mo

Tc

Ru

Rh

Pd

Ag

Cd

In

Sn

Sb

Te

I

Xe

Cs

Ba

La

Ce

Pr

Nd

Pm

Sm

Eu

Gd

Tb

Dy

Ho

Er

Tm

Yb

Lu

Hf

Ta

W

Re

Os

Ir

Pt

Au

Hg

Tl

Pb

Bi

Po

At

Rn

Fr

Ra

Ac

Th

Pa

U

Np

Pu

Am

Cm

Bk

Cf

Es

Fm

Md

No

Lr

Rf

Db

Sg

Bh

Hs

Mt

Ds

Rg

Cn

Uut

Fl

Uup

Lv

Uus

Uuo

type AtomLabel = String

An atom label. They may come from pdb or force field parameter files or can be assigned by other ways just to distinguish specific atoms.

type FFType = String

These are labels for molecular mechanics software. The strings are basically arbitrary and depending on the MM software used.

data Atom

An Atom in a Molecule.

Constructors

Atom

Fields

_atom_Index :: Int

The index of an atom in a molecule. For counting conventions refer to Molecule and IndType.

_atom_Element :: Element

Chemical Element of the atom.

_atom_Label :: AtomLabel

Label, e.g. from a pdb, just for identification, can be empty.

_atom_IsPseudo :: Bool

Boolean, telling if this is a pseudo atom, introduced because a bond was broken.

_atom_FFType :: FFType

Label depending on the MM software used, identifying topological atom.

_atom_PCharge :: Maybe Double

Possibly a partial charge.

_atom_Coordinates :: Vector Double

Coordinates of the atom, cartesian in R³. Relies on the parser to fill with exactly 3 values.

atom_PCharge :: Lens' Atom (Maybe Double)

atom_Label :: Lens' Atom AtomLabel

atom_IsPseudo :: Lens' Atom Bool

atom_Index :: Lens' Atom Int

atom_FFType :: Lens' Atom FFType

atom_Element :: Lens' Atom Element

atom_Coordinates :: Lens' Atom (Vector Double)

data Molecule

A molecule, which might be the whole system, an ONIOM layer or a fragment of the system, each containing possibly even higher layers for ONIOM or fragments. Stores all associated informations of a layer.

Starting from a top level molecule, all atoms and bonds of the system are expected to be in the in this top layer (except pseudoatoms of deeper layers). Therefore if atoms are in a deeper layers of the recursion, their information is not used to describe a higher layer. Instead, all atoms of deeper layers (except pseudoatoms) must be also replicated in a higher layer.

Two possible types of index countings are possible:

Global: All atoms in the system (which should be present at the top layer molecule) except pseudoatoms are counted in the top layer. The fragments might therefore not start at atom with index 0, but depending on where they are in a higher level with higher numbers. Pseudoatoms are counted after all other atoms are counted.

Individual: Counting happens for each molecule individually. In this counting scheme, pseudoatoms do not need any special treatment.

Constructors

Molecule

Fields

_molecule_Label :: String

Comment or identifier of a molecule. Can be empty.

_molecule_Atoms :: Vector Atom

A Vector of Atoms.

_molecule_Bonds :: IntMap IntSet

An IntMap, mapping the index of an Atom in the Molecule to the indices of all Atoms, to which it binds.

_molecule_SubMol :: Vector Molecule

A Molecule might contain other molecules. These might be fragments or higher level ONIOM layers.

_molecule_IndType :: IndType

Which counting scheme for atoms is used in this representation. Must be same for all layers.

_molecule_Energy :: Maybe Double

An energy, that might have been calculated.

_molecule_Gradient :: Maybe (Vector Double)

A gradient, that might have been calculated.

_molecule_Hessian :: Maybe (UMatrix Double)

A hessian, that might have been calculated.

molecule_SubMol :: Lens' Molecule (Vector Molecule)

molecule_Label :: Lens' Molecule String

molecule_IndType :: Lens' Molecule IndType

molecule_Hessian :: Lens' Molecule (Maybe (UMatrix Double))

molecule_Gradient :: Lens' Molecule (Maybe (Vector Double))

molecule_Energy :: Lens' Molecule (Maybe Double)

molecule_Bonds :: Lens' Molecule (IntMap IntSet)

molecule_Atoms :: Lens' Molecule (Vector Atom)

type Trajectory = Vector Molecule

Trajectories are simply a vector of Molecules.

Data types to describe the calculation niveau. They are hierarchically organised and try to cover everything, so that Spicy is fully aware of the calculation niveau. This has the advantage, that users basically do not need to know the input of a specific program to perform a Spicy calculation.