Spicy.UnitTests.Data

Description

The module collects data only used for unit testing, such as the correct answers for unit tests.

Documentation

moleculeEmpty :: Molecule

Empty molecule.

moleculeHFeCNxH2O :: Molecule

Test Molecule with all information a Molecule could have.

moleculeHFeCNxH2OTXYZ :: Molecule

Test Molecule with all informations a TXYZ could have (without calling Tinker).

moleculeHFeCNxH2OXYZ :: Molecule

Test Molecule with all informations a XYZ could have.

moleculeHFeCNxH2OMOL2 :: Molecule

Test Molecule with all informations a TXYZ could have.

textHFeCNxH2OTXYZ :: Text

Test Molecule as TXYZ file.

testHFeCNxH2OXYZ :: Text

Test Molecule as XYZ file.

testHFeCNxH2OMOL2 :: Text

Test Molecule as MOL2 file.

testHFeCNxH2OSpicy :: Text

Test Molecule as Spicy file.