Spicy.Math

Description

This module defines basic algebraic operations used throughout the program. Numerical heavy and most other operations are implemented using Accelerate, to provide parallel operations. Note that all runQ provided functions must be typed without typeclasses but by concrete types.

The operations here accept some insecurities (like not checking if both vectors of a dot product have equal lenght) and trust the caller.

Documentation

(<.>) :: Num a => Seq a -> Seq a -> a

Dot product of two Seqs.

vLength :: Floating a => Seq a -> a

Length of a Seq.

vDistance :: Floating a => Seq a -> Seq a -> a

Distance between 2 points (Seqs).

vAngle :: Floating a => Seq a -> Seq a -> a

Angle in radian between 2 Seqs.

vCross :: Seq Double -> Seq Double -> Either String (Seq Double)

3D cross product of 2 Seqs.