Spicy.MolWriter

Description

A module which converts the internal Molecule representation to a string, which is a common chemical file format, that can be read by Avogadro, VMD, OpenBabel etc.. The writers are not fool proof with respect to force field types, which should always be remembered when usings its results.

Documentation

writeXYZ :: Molecule -> Text

Write a Molden XYZ file from a molecule. This format ignores all deep level layers of a molecule.

writeTXYZ :: Molecule -> Text

Write a Tinker XYZ from a Molecule. The writer trusts the _atom_FFType to be correct (if set) and will simply write them out. Therefore it is possible, that wrong atom types can be written. If they are not set, the writer will simply equalise all atom types to 0, which is OK for visualisation but obviously not for MM.

This format ingores all deeper level layers of a molecule.

writeMOL2 :: Molecule -> Text

Write a simplified .mol2 file (Tripos SYBYL) from a Molecule, containing the atoms, connectivities (single bonds only) and partial charges. The writer is not fool proof and will happily accept any _atom_FFType that is supplied, even if it is not a TRIPOS SYBYL atom type. This can lead to mol2 files that have correct topology and geometry, but visualisation programs wont be able to assign correct elements.

writePDB :: Molecule -> Text

writeSpicy :: Molecule -> Text

Write Spicy format, which is an AESON generated JSON document, directly representing the data type of Molecule.