Spicy.Molecule.Util

Description

This module provides functions to manipulate basic data structures of Molecules, such as indexing.

Documentation

checkMolecule :: Molecule -> Either String Molecule

Check if a single Molecule layer has sane indices (Key) in the IntMap for _molecule_Atoms and _molecule_SubMol, meaning the bonds do not include Atoms, that do not exist. Then recursively check if the Atoms of deeper layers are proper subsets of the higher layer, both for _molecule_Atoms and _molecule_Bonds. Also checks if fragments of the same layer are disjoint. Returns the Molecule without modifications if it is sane, a string with description of the found problem otherwise. This check is exhaustive and potentially expensive.

reIndexMolecule :: IntMap Int -> Molecule -> Either String Molecule

Reindex a complete Molecule, including all its deeper layers in _molecule_SubMol) by mappings from a global replacement Map, mapping old to new indices. This function assumes, that your molecule is sane in the overall assumptions of this program. This means that the lower layers obey the counting scheme of the atoms of the higher layers and pseudoatoms come last.

groupTupleSeq :: Seq (Int, Int) -> IntMap IntSet

Group by the first tuple element and within this group build an IntSet of the the second tuple elements.

groupBy :: (a -> a -> Bool) -> Seq a -> Seq (Seq a)

This function implements [groupBy](http:/hackage.haskell.orgpackagebase-4.12.0.0docs/Data-List.html#v:groupBy) as in Data.List: "The group function takes a list and returns a list of lists such that the concatenation of the result is equal to the argument. Moreover, each sublist in the result contains only equal elements."